requirements.txt
python-dotenv
faiss-cpu
pydantic
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.100.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain[openai]
langchain[google-vertexai]
langchain-text-splitters
langchain
langchain-community
langchain-core==0.1.52
langsmith==0.1.0
langchain-openai
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.6
packaging==23.2
pandas==2.1.4
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
uvicorn[standard]==0.23.2